// ui/src/lib/booking-hotel-api/model/bed-type-model.ts
/**
 * main-app API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BedTypeModel { 
    id: number;
    name: string;
    iconRef: string;
    createdAt: string;
    updatedAt: string;
    removed: boolean;
}